Regenerate wrapper
Wrapper was generated with the older version of the plugin (3.2.0). If we update the wrapper plugin version, we probably should also re-generate the wrapper itself (otherwise, what would be the point of bumping the plugin version?)

The default distribution type has changed, and now it is a script-only (https://maven.apache.org/tools/wrapper/#Usage_with_or_without_Binary_JAR) which is "nice" since it ensures that there are no binaries included in the sources.